refactored preprocessor handling.
