+++ src/input/original-source.ts
