+++ src/rules-engine/registry/rule-registry.ts
@@ -10,8 +10,8 @@ import { K8sManifest } from '../../manifests/k8s-manifest';
 import { ManifestLoader } from '../../manifests/manifests-loader';
 import { ManifestPackage } from '../../manifests/manifest-package';
 import { KubeviousKinds, KUBEVIOUS_API_NAME } from '../../types/kubevious';
-import { InputSource } from '../../manifests/input-source';
-import { OriginalSource } from '../../manifests/original-source';
+import { InputSource } from '../../input/input-source';
+import { OriginalSource } from '../../input/original-source';
 
 
 export interface RuleRegistryLoadOptions {
